Fixed Vuetable pagination issues

+++ src/controllers/id/CmsLicensesController.php
@@ -132,7 +132,7 @@ class CmsLicensesController extends BaseController
 
         $filter = Craft::$app->getRequest()->getParam('filter');
         $limit = Craft::$app->getRequest()->getParam('limit', 10);
-        $page = Craft::$app->getRequest()->getParam('page', 1);
+        $page = (int) Craft::$app->getRequest()->getParam('page', 1);
         $orderBy = Craft::$app->getRequest()->getParam('orderBy');
         $ascending = Craft::$app->getRequest()->getParam('ascending');
         $byColumn = Craft::$app->getRequest()->getParam('byColumn');
